Strategy

+++ src/main/java/Interfaces/Formula.java
@@ -3,5 +3,5 @@ package Interfaces;
 public interface Formula
 {
     String asString();
-    float asValue();
+    Double asValue();
 }
